Fixes #62. Simple way to specify dependency options
Configuration has a new QLineEdit that takes a string of the form

<pkg>:<option>=<value>[,<repeat>]

so that option values for dependencies can be incorporated into the build.

The QLineEdit has a validator to ensure the format.

Conan commands that take options, as well as the lockfile generation as part
of recipe loading, will automatically use additional options.

This is a (hopefully) short term solution as 'full options' are not yet
expressed as part of the UI.