Fixes #46. Submenu parentage should be to the parent menu
This was giving a warning on Weyland if the parent was not set correctly.

WARNING:cruiz.entrypoint:setGrabPopup called with a parent, QtWaylandClient::QWaylandXdgSurface(0x55c528e78d30) which does not match the current topmost grabbing popup, QtWaylandClient::QWaylandXdgSurface(0x55c528e91760) According to the xdg-shell protocol, this is not allowed. The wayland QPA plugin is currently handling it by setting the parent to the topmost grabbing popup. Note, however, that this may cause positioning errors and popups closing unxpectedly because xdg-shell mandate that child popups close before parents