Fixes #146 Determine Conan location via import mechanisms rather than the PATH
For example, using pipx, the conan executable may be visible via the PATH
but is not the imported package that cruiz uses.

cruiz no longer assumes it can see the conan executable, which was never used
to execute commands anyway, and instead just uses conan as an imported package.

Wrote the arbitrary command in Conan2 to use the usual entry point for the
conan package, to avoid calling the executable.