Fixes #187 Source and binary builds no longer delete the generated files
The main reason for this is so that pip editable installs now have all
the same files at application startup that a regular install does.

To address a warning, writing the RELEASE_VERSION.py now ends with a new line.

tox linting now performs on editable installs again, because we want it to operate
on the files in the working directory.