Fixes #217 Move tested code to cruizlib
This first introduces the separate cruizlib, which is not to contain any GUI aspects.

Only the ConanHook dataclass is currently tested, so move this over to cruizlib.

mypy and pylint check for any parts of the existing code imports that need changing.
mypy and pylint check _more_ for cruizlib since this is new.