Tests: Exercise meta available hooks
Some code re-use for the hook path can be used here.

Also added a fixture to add a hook, as otherwise some of the code to find available hooks is not exercised.
This fixture name starts with an underscore, because it does not need referencing in the use here, which causes a pylint unused argument warning.