Tests: Exercise conan lock create
Some adjustments needed to be made for Conan 1.17.1 to work. Calling conan lock create predates working with that version of Conan.

There is also a known bug in Conan 2's lock create code in cruiz, which has an issue, but isn't fixed here. The bug was caught by the tests though, so is an expected failure now.