Tests: As recommended, switch pytest to importlib mode
As sys.path is no longer modified, moving the test utilities to another folder, which
pytest can be told to search for.

Note that on Python 3.10 in VSCode, this is a problem for test discovery, as it cannot
find the testexceptions module.
A workaround for that is to set PYTHONPATH before launching VSCode to the directory containing it.