Refactor: Convert all tests to use multiprocessing queues
Why is this happening?

cruiz works in a multiprocessing way. And while single process tests are easier to work with
to test more of cruiz, such as core functionality of message processing which checks for
Conan symbols being brought over, this cannot be done in a single process, because Conan is
imported by the workers, so the workers cannot be run in the same process.