Refactor: Watcher threads must explicitly end with an End message.
Otherwise, the loop can wait indefinitely on a queue.get() if a Success
is sent, and then some logging is sent after, such as in the time commands option.
This was because the logging skipped the end of the loop test. It couldn't always
do it, because you didn't know if more was coming.